ADD: messages in the rooms

+++ base/views.py
@@ -1,10 +1,14 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
 from .forms import RoomForm
-from .models import Room, Topic
+from .models import Room, Topic, Message
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from django.contrib import messages
+
+from django.contrib.auth.forms import UserCreationForm
 
 
 # rooms = [
@@ -16,8 +20,13 @@ from django.contrib import messages
 # Create your views here.
 
 def loginPage(request):
+  page = 'login'
+
+  if request.user.is_authenticated:
+    return redirect('home')
+
   if request.method == "POST":
-    username = request.POST.get('username')
+    username = request.POST.get('username').lower()
     password = request.POST.get('password')
 
     try:
@@ -33,28 +42,58 @@ def loginPage(request):
       messages.error(request, 'Username or password does not exist')
 
 
-  context = {}
+  context = {'page' : page}
   return render(request, 'base/login__register.html', context)
 
 def logoutUser(request):
   logout(request)
   return redirect('home')
 
+def registerPage(request):
+  form = UserCreationForm()
+
+  if request.method == 'POST':
+    form = UserCreationForm(request.POST)
+    if form.is_valid():
+      user = form.save(commit=False)
+      user.username = user.username.lower()
+      user.save()
+      login(request, user)
+      return redirect('home')
+    else:
+      messages.error(request, 'Some error occurred during registration!')
+  context = {'form' : form}
+  return render(request, 'base/login__register.html', context)
+
+
 def home(request):
   q = request.GET.get('q') if request.GET.get('q') != None else ''
   topics = Topic.objects.all()
   rooms = Room.objects.filter(Q(topico__name__icontains = q) | Q(name__icontains = q) | Q(description__icontains = q))
-
+  room_message = Message.objects.filter(Q(room__topico__name__icontains = q))
   room_count = rooms.count()
 
-  context =  {'rooms' : rooms, 'topics' : topics, 'room_count' : room_count}
+  context =  {'rooms' : rooms, 'topics' : topics, 'room_count' : room_count, 'room_messages': room_message}
   return render(request,'base/home.html',context)
 
 def room(request, pk):
   room = Room.objects.get(id = pk)
-  context = {'room' : room}
+  participants = room.participants.all()
+  room_messages = room.message_set.all().order_by('-create')
+
+  if request.method == 'POST':
+    message = Message.objects.create(
+       user = request.user,
+       room = room,
+       body = request.POST.get('body')
+    )
+    room.participants.add(request.user)
+    return redirect('room', pk = room.id)
+
+  context = {'room' : room, 'room_messages' : room_messages, 'participants' : participants}
   return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
   form = RoomForm()
   if request.method == 'POST':
@@ -65,9 +104,13 @@ def createRoom(request):
   context = {'form':form}
   return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
   room = Room.objects.get(id = pk)
   form = RoomForm(instance=room)
+
+  if request.user != room.host:
+    return HttpResponse("Your aren't allowed here!")
 
   if request.method == 'POST':
     form = RoomForm(request.POST, instance=room)
@@ -78,10 +121,29 @@ def updateRoom(request, pk):
   context = { 'form' : form}
   return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
   room = Room.objects.get(id = pk)
+
+  if request.user != room.host:
+    return HttpResponse("Your aren't allowed here!")
+
   if request.method == "POST":
     room.delete()
     return redirect('home')
 
   return render(request, 'base/delete.html', {'obj' : room})
+
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+  message = Message.objects.get(id = pk)
+
+  if request.user != message.user:
+    return HttpResponse("Your aren't allowed here!")
+
+  if request.method == "POST":
+    message.delete()
+    return redirect('home')
+
+  return render(request, 'base/delete.html', {'obj' : message})
